Fin(sorta)
Tripster Whipster, where is my mixer
My heart is a fix and my feet are a hipster my hand is a gong and my head is too long and means of the mean mean are already gone
So whit is to do but to follow the true,
the strawberry hallowmerry
fallow-wary stew?
The Men who don't swear of the carmely stare and who jump up the stairs to the stars of the bear.
Polaris I swear this until my mind sing of the ickel no spickel of pineberry swing
I love the Formalo Contalo tudriz
tith mixer noblitzer Politzers Politz.
And when the stars meet on the Maple Leaf  Ground the small tiny whispers are sent to the pound bought by the pound, heard to resound, and built by the mound.
Tate. Tape. Hape. Gape. Vape.